+++ src/platform_impl/ios/window.rs
@@ -196,7 +196,7 @@ impl Inner {
         Err(ExternalError::NotSupported(NotSupportedError::new()))
     }
 
-    pub fn set_ignore_mouse_events(&self, _ignore: bool) -> Result<(), ExternalError> {
+    pub fn set_cursor_hittest(&self, _hittest: bool) -> Result<(), ExternalError> {
         Err(ExternalError::NotSupported(NotSupportedError::new()))
     }