+++ src/platform_impl/linux/mod.rs
@@ -397,8 +397,8 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_ignore_mouse_events(&self, _ignore: bool) -> Result<(), ExternalError> {
-        Err(ExternalError::NotSupported(NotSupportedError::new()))
+    pub fn set_cursor_hittest(&self, hittest: bool) -> Result<(), ExternalError> {
+        x11_or_wayland!(match self; Window(w) => w.set_cursor_hittest(hittest))
     }
 
     #[inline]